requirements.txt
cmake
numpy==1.18.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
pandas==1.0.3
Pillow==8.0.1
psycopg2==2.8.4
SQLAlchemy==1.3.19
tensorflow-cpu
Keras==2.3.1
opencv-python==4.4.0.44
